+++ custom_components/rfplayer/__init__.py
@@ -1,14 +1,13 @@
 """Support for Rfplayer devices."""
-import asyncio
-from collections import defaultdict
+
+from asyncio import BaseTransport, timeout
 import copy
 import logging
+from typing import Any
 
-import async_timeout
-from serial import SerialException
-from homeassistant.util import slugify
 import voluptuous as vol
 
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     ATTR_STATE,
@@ -19,15 +18,15 @@ from homeassistant.const import (
     CONF_PROTOCOL,
     EVENT_HOMEASSISTANT_STOP,
 )
-from homeassistant.core import CoreState, callback
+from homeassistant.core import CoreState, HomeAssistant, callback
 import homeassistant.helpers.config_validation as cv
+from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.dispatcher import (
     async_dispatcher_connect,
     async_dispatcher_send,
 )
-from homeassistant.helpers import device_registry as dr
-from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.restore_state import RestoreEntity
+from homeassistant.util import slugify
 
 from .const import (
     CONF_AUTOMATIC_ADD,
@@ -40,6 +39,7 @@ from .const import (
     EVENT_BUTTON_PRESSED,
     EVENT_KEY_COMMAND,
     EVENT_KEY_ID,
+    EVENT_KEY_JAMMING,
     EVENT_KEY_SENSOR,
     PLATFORMS,
     RFPLAYER_PROTOCOL,
@@ -48,7 +48,11 @@ from .const import (
     SIGNAL_EVENT,
     SIGNAL_HANDLE_EVENT,
 )
-from .rflib.rfpprotocol import create_rfplayer_connection
+from .rflib.rfpprotocol import (
+    ProtocolBase,
+    RfplayerProtocol,
+    create_rfplayer_connection,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -59,7 +63,7 @@ SEND_COMMAND_SCHEMA = vol.Schema(
         vol.Required(CONF_COMMAND): cv.string,
         vol.Optional(CONF_DEVICE_ADDRESS): cv.string,
         vol.Optional(CONF_DEVICE_ID): cv.string,
-        vol.Required(CONF_AUTOMATIC_ADD, default=False): cv.boolean,
+        vol.Required(CONF_AUTOMATIC_ADD, default=False): cv.boolean,  # type: ignore
     }
 )
 
@@ -69,6 +73,8 @@ def identify_event_type(event):
 
     Async friendly.
     """
+    if "JAMMING_" in event.get(EVENT_KEY_ID):
+        return EVENT_KEY_JAMMING
     if EVENT_KEY_COMMAND in event:
         return EVENT_KEY_COMMAND
     if EVENT_KEY_SENSOR in event:
@@ -76,10 +82,25 @@ def identify_event_type(event):
     return "unknown"
 
 
-async def async_setup_entry(hass, entry):
+# pylint: disable-next=too-many-statements
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up GCE RFPlayer from a config entry."""
+
     config = entry.data
     options = entry.options
+
+    hass.data.setdefault(
+        DOMAIN,
+        {
+            CONF_DEVICE: config[CONF_DEVICE],
+            DATA_ENTITY_LOOKUP: {
+                EVENT_KEY_COMMAND: {},
+                EVENT_KEY_SENSOR: {},
+                EVENT_KEY_JAMMING: {},
+            },
+            DATA_DEVICE_REGISTER: {},
+        },
+    )
 
     async def async_send_command(call):
         """Send Rfplayer command."""
@@ -96,7 +117,9 @@ async def async_setup_entry(hass, entry):
             event_id = "_".join(
                 [
                     call.data[CONF_PROTOCOL],
-                    call.data.get(CONF_DEVICE_ID) or call.data.get(CONF_DEVICE_ADDRESS),
+                    call.data.get(CONF_DEVICE_ID)
+                    or call.data.get(CONF_DEVICE_ADDRESS)
+                    or call.data.get(CONF_COMMAND),
                 ]
             )
             device = {
@@ -132,23 +155,21 @@ async def async_setup_entry(hass, entry):
         # Lookup entities who registered this device id as device id or alias
         event_id = event.get(EVENT_KEY_ID)
 
-        entity_id = hass.data[DOMAIN][DATA_ENTITY_LOOKUP][event_type][event_id]
+        entity_id = hass.data[DOMAIN][DATA_ENTITY_LOOKUP][event_type].get(event_id)
 
         if entity_id:
             # Propagate event to every entity matching the device id
             _LOGGER.debug("passing event to %s", entity_id)
             async_dispatcher_send(hass, SIGNAL_HANDLE_EVENT.format(entity_id), event)
+        elif event_type in hass.data[DOMAIN][DATA_DEVICE_REGISTER]:
+            _LOGGER.debug("event_id not known, adding new device")
+            hass.data[DOMAIN][DATA_ENTITY_LOOKUP][event_type][event_id] = event
+            _add_device_to_base_config(event, event_id)
+            hass.async_create_task(
+                hass.data[DOMAIN][DATA_DEVICE_REGISTER][event_type](event)
+            )
         else:
-            # If device is not yet known, register with platform (if loaded)
-            if event_type in hass.data[DOMAIN][DATA_DEVICE_REGISTER]:
-                _LOGGER.debug("device_id not known, adding new device")
-                hass.data[DOMAIN][DATA_ENTITY_LOOKUP][event_type][event_id] = event
-                _add_device_to_base_config(event, event_id)
-                hass.async_create_task(
-                    hass.data[DOMAIN][DATA_DEVICE_REGISTER][event_type](event)
-                )
-            else:
-                _LOGGER.debug("device_id not known and automatic add disabled")
+            _LOGGER.debug("event_id not known and automatic add disabled")
 
     @callback
     def _add_device_to_base_config(event, event_id):
@@ -162,6 +183,8 @@ async def async_setup_entry(hass, entry):
     def reconnect(exc=None):
         """Schedule reconnect after connection has been unexpectedly lost."""
         # Reset protocol binding before starting reconnect
+        if exc:
+            _LOGGER.error(exc)
         hass.data[DOMAIN][RFPLAYER_PROTOCOL] = None
 
         async_dispatcher_send(hass, SIGNAL_AVAILABILITY, False)
@@ -180,19 +203,18 @@ async def async_setup_entry(hass, entry):
             disconnect_callback=reconnect,
             loop=hass.loop,
         )
+        transport: BaseTransport
+        protocol: ProtocolBase
 
         try:
-            with async_timeout.timeout(CONNECTION_TIMEOUT):
+            async with timeout(CONNECTION_TIMEOUT):
                 transport, protocol = await connection
 
-        except (
-            SerialException,
-            OSError,
-            asyncio.TimeoutError,
-        ) as exc:
+        except (TimeoutError, OSError) as exc:
             reconnect_interval = config[CONF_RECONNECT_INTERVAL]
             _LOGGER.exception(
-                "Error connecting to Rfplayer, reconnecting in %s", reconnect_interval
+                "Error connecting to Rfplayer, reconnecting in %s",
+                reconnect_interval,
             )
             # Connection to Rfplayer device is lost, make entities unavailable
             async_dispatcher_send(hass, SIGNAL_AVAILABILITY, False)
@@ -204,15 +226,9 @@ async def async_setup_entry(hass, entry):
         # mark entities as available
         async_dispatcher_send(hass, SIGNAL_AVAILABILITY, True)
 
-        hass.data[DOMAIN] = {
-            RFPLAYER_PROTOCOL: protocol,
-            CONF_DEVICE: config[CONF_DEVICE],
-            DATA_ENTITY_LOOKUP: {
-                EVENT_KEY_COMMAND: defaultdict(list),
-                EVENT_KEY_SENSOR: defaultdict(list),
-            },
-            DATA_DEVICE_REGISTER: {},
-        }
+        hass.data[DOMAIN][RFPLAYER_PROTOCOL] = protocol
+
+        entry.add_update_listener(_async_update_listener)
 
         if options.get(CONF_AUTOMATIC_ADD, config[CONF_AUTOMATIC_ADD]) is True:
             for device_type in "sensor", "command":
@@ -234,42 +250,72 @@ async def async_setup_entry(hass, entry):
     return True
 
 
+async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Unload a config entry."""
+    await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
+
+    hass.data.pop(DOMAIN)
+
+    return True
+
+
+async def _async_update_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
+    """Handle options update."""
+    await hass.config_entries.async_reload(entry.entry_id)
+
+
+# pylint: disable-next=too-many-instance-attributes
 class RfplayerDevice(RestoreEntity):
     """Representation of a Rfplayer device.
 
     Contains the common logic for Rfplayer entities.
     """
 
-    platform = None
-    _state = None
     _available = True
+    _attr_assumed_state = True
+    _attr_should_poll = False
 
+    # pylint: disable-next=too-many-arguments
     def __init__(
         self,
-        protocol,
-        device_address=None,
-        device_id=None,
-        initial_event=None,
-        name=None,
-    ):
+        protocol: str,
+        device_address: str | None = None,
+        device_id: str | None = None,
+        initial_event: dict[str, Any] | None = None,
+        name: str | None = None,
+        unique_id: str | None = None,
+    ) -> None:
         """Initialize the device."""
-        # Rflink specific attributes for every component type
+        # Rfplayer specific attributes for every component type
+        if not (name or device_address or device_id) or not (
+            unique_id or device_address or device_id
+        ):
+            raise TypeError("Incorrect arguments for RfplayerDevice")
         self._initial_event = initial_event
         self._protocol = protocol
         self._device_id = device_id
         self._device_address = device_address
         self._event = None
-        self._state: bool = None
-        self._attr_assumed_state = True
-        if name is not None:
-            self._attr_name = name
-            self._attr_unique_id = slugify(f"{protocol}_{name}")
-        else:
-            self._attr_name = f"{protocol} {device_id or device_address}"
-            self._attr_unique_id = slugify(f"{protocol}_{device_id or device_address}")
 
-    async def _async_send_command(self, command, *args):
-        rfplayer = self.hass.data[DOMAIN][RFPLAYER_PROTOCOL]
+        self._attr_name = name or f"{protocol} {device_address or device_id}"
+        self._attr_unique_id = slugify(
+            "_".join(
+                [protocol, unique_id or device_address or device_id]  # type: ignore[list-item]
+            )
+        )
+
+    @property
+    def device_info(self) -> DeviceInfo:
+        """Return device registry information for this entity."""
+        return DeviceInfo(
+            identifiers={(DOMAIN, self.hass.data[DOMAIN][CONF_DEVICE])},
+            manufacturer="GCE",
+            model="RFPlayer",
+            name="RFPlayer",
+        )
+
+    async def _async_send_command(self, command) -> None:
+        rfplayer: RfplayerProtocol = self.hass.data[DOMAIN][RFPLAYER_PROTOCOL]
         await rfplayer.send_command_ack(
             command=command,
             protocol=self._protocol,
@@ -278,7 +324,7 @@ class RfplayerDevice(RestoreEntity):
         )
 
     @callback
-    def handle_event_callback(self, event):
+    def handle_event_callback(self, event: dict[str, Any]) -> None:
         """Handle incoming event for device type."""
         # Call platform specific event handler
         self._handle_event(event)
@@ -296,46 +342,23 @@ class RfplayerDevice(RestoreEntity):
                 "Fired bus event for %s: %s", self.entity_id, event[EVENT_KEY_COMMAND]
             )
 
-    def _handle_event(self, event):
+    def _handle_event(self, event) -> None:
         """Platform specific event handler."""
-        raise NotImplementedError()
+        raise NotImplementedError
 
     @property
-    def should_poll(self):
-        """No polling needed."""
-        return False
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device registry information for this entity."""
-        return DeviceInfo(
-            identifiers={
-                (
-                    DOMAIN,
-                    self.hass.data[DOMAIN][
-                        CONF_DEVICE
-                    ],  # + "_" + self._attr_unique_id,
-                )
-            },
-            manufacturer="GCE",
-            model="RFPlayer",
-            name="RFPlayer",
-        )
-
-    @property
-    def available(self):
+    def available(self) -> bool:
         """Return True if entity is available."""
         return bool(self._protocol)
 
     @callback
-    def _availability_callback(self, availability):
+    def _availability_callback(self, availability) -> None:
         """Update availability state."""
         self._available = availability
         self.async_write_ha_state()
 
-    async def async_added_to_hass(self):
+    async def async_added_to_hass(self) -> None:
         """Register update callback."""
-        await super().async_added_to_hass()
         self.async_on_remove(
             async_dispatcher_connect(
                 self.hass, SIGNAL_AVAILABILITY, self._availability_callback
@@ -352,13 +375,3 @@ class RfplayerDevice(RestoreEntity):
         # Process the initial event now that the entity is created
         if self._initial_event:
             self.handle_event_callback(self._initial_event)
-
-    async def async_will_remove_from_hass(self):
-        """Clean when entity removed."""
-        await super().async_will_remove_from_hass()
-        device_registry = dr.async_get(self.hass)
-        device = device_registry.async_get_device(
-            (DOMAIN, self.hass.data[DOMAIN][CONF_DEVICE] + "_" + self._attr_unique_id)
-        )
-        if device:
-            device_registry.async_remove_device(device)