+++ custom_components/rfplayer/rflib/exception.py
@@ -1,0 +1,5 @@
+"""RfPlayer exceptions."""
+
+
+class RfPlayerException(Exception):
+    """Generic RfPlayer exception."""