+++ custom_components/rfplayer/switch.py
@@ -1,14 +1,19 @@
 """Support for Rfplayer switch."""
+
 import logging
+from typing import Any
 
 from homeassistant.components.switch import SwitchEntity
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_DEVICE_ID, CONF_DEVICES, CONF_PROTOCOL
-from homeassistant.core import callback
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import DATA_DEVICE_REGISTER, EVENT_KEY_COMMAND, RfplayerDevice
 from .const import (
     COMMAND_OFF,
     COMMAND_ON,
+    CONF_AUTOMATIC_ADD,
     CONF_DEVICE_ADDRESS,
     DATA_ENTITY_LOOKUP,
     DOMAIN,
@@ -18,7 +23,9 @@ from .const import (
 _LOGGER = logging.getLogger(__name__)
 
 
-async def async_setup_entry(hass, entry, async_add_entities):
+async def async_setup_entry(
+    hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
+) -> None:
     """Set up the Rfplayer platform."""
     config = entry.data
     options = entry.options
@@ -36,23 +43,38 @@ async def async_setup_entry(hass, entry, async_add_entities):
         async_add_entities([device])
 
     if CONF_DEVICES in config:
-        for device_id, device_info in config[CONF_DEVICES].items():
-            if EVENT_KEY_COMMAND in device_info:
-                await add_new_device(device_info)
+        for device in config[CONF_DEVICES].values():
+            if EVENT_KEY_COMMAND in device:
+                await add_new_device(device)
 
-    hass.data[DOMAIN][DATA_DEVICE_REGISTER][EVENT_KEY_COMMAND] = add_new_device
+    if options.get(CONF_AUTOMATIC_ADD, config[CONF_AUTOMATIC_ADD]):
+        hass.data[DOMAIN][DATA_DEVICE_REGISTER][EVENT_KEY_COMMAND] = add_new_device
 
 
 class RfplayerSwitch(RfplayerDevice, SwitchEntity):
     """Representation of a Rfplayer sensor."""
 
-    async def async_added_to_hass(self):
+    # pylint: disable-next=too-many-arguments
+    def __init__(
+        self,
+        protocol: str,
+        device_id: str | None = None,
+        device_address: str | None = None,
+        initial_event: dict[str, Any] | None = None,
+        name: str | None = None,
+    ) -> None:
+        """Handle switch specific args and super init."""
+        self._state: bool | None = None
+        super().__init__(protocol, device_address, device_id, initial_event, name)
+
+    async def async_added_to_hass(self) -> None:
         """Restore RFPlayer device state (ON/OFF)."""
         await super().async_added_to_hass()
 
-        self.hass.data[DOMAIN][DATA_ENTITY_LOOKUP][EVENT_KEY_COMMAND][
-            self._initial_event[EVENT_KEY_ID]
-        ] = self.entity_id
+        if self._initial_event:
+            self.hass.data[DOMAIN][DATA_ENTITY_LOOKUP][EVENT_KEY_COMMAND][
+                self._initial_event[EVENT_KEY_ID]
+            ] = self.entity_id
 
         if self._event is None:
             old_state = await self.async_get_last_state()
@@ -68,17 +90,17 @@ class RfplayerSwitch(RfplayerDevice, SwitchEntity):
             self._state = False
 
     @property
-    def is_on(self):
+    def is_on(self) -> bool | None:
         """Return true if device is on."""
         return self._state
 
-    async def async_turn_on(self, **kwargs):
+    async def async_turn_on(self, **kwargs: Any) -> None:
         """Turn the device on."""
         await self._async_send_command(COMMAND_ON)
         self._state = True
         self.async_write_ha_state()
 
-    async def async_turn_off(self, **kwargs):
+    async def async_turn_off(self, **kwargs: Any) -> None:
         """Turn the device off."""
         await self._async_send_command(COMMAND_OFF)
         self._state = False