+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 0.7.0
+
+- Fix default conf missing error during initialization
+- Lint and code improvments
+- Bump pyserial-asyncio
+- Add service translation
+- Handle no USB device found during init
+- Replace the jamming sensor by a binary_sensor and add events catching to it
+- Allow to specify manual path to USB device
+- Disallow multi instance of the integration
+- Fix options changes not reloading integration
+
 ## 0.6.1
 
 - Fix incapacity to create switch entity when automatic entity add is disabled at bootstrap